double to int for predict function perceptron

+++ src/header/perceptron.hpp
@@ -7,7 +7,7 @@ class Perceptron
 public:
     Perceptron(std::vector<double> weights, double bias, double learningRate);
 
-    double predict(const std::vector<int>& x) const;
+    int predict(const std::vector<int>& x) const;
     void train(const std::vector<std::vector<int>>& inputs, const std::vector<int>& targets, int epochs);
     void __str__(int verbose) const;
